+++ kpops/components/streams_bootstrap/model.py
@@ -9,7 +9,6 @@ from kpops.components.base_components.helm_app import HelmAppValues
 from kpops.components.common.kubernetes_model import (
     ImagePullPolicy,
     ProtocolSchema,
-    ResourceDefinition,
     Resources,
     ServiceType,
     Toleration,
@@ -43,8 +42,9 @@ class PortConfig(CamelCaseConfigModel, DescConfigModel):
         default=None,
         description=describe_attr("name", __doc__),
     )
-    schema: ProtocolSchema = Field(
+    schema_: ProtocolSchema = Field(
         default=ProtocolSchema.TCP,
+        alias="schema",  # because schema is already a builtin of Pydantic
         description=describe_attr("schema", __doc__),
     )
     service_port: int | None = Field(
@@ -69,8 +69,8 @@ class ServiceConfig(CamelCaseConfigModel, DescConfigModel):
         default_factory=dict,
         description=describe_attr("labels", __doc__),
     )
-    type: ServiceType = Field(
-        default=ServiceType.CLUSTER_IP,
+    type: ServiceType | None = Field(
+        default=None,
         description=describe_attr("type", __doc__),
     )
 
@@ -82,8 +82,8 @@ class JavaOptions(CamelCaseConfigModel, DescConfigModel):
     :param others: List of Java VM options passed to the streams app.
     """
 
-    max_RAM_percentage: int = Field(
-        default=75,
+    max_RAM_percentage: int | None = Field(
+        default=None,
         description=describe_attr("max_RAM_percentage", __doc__),
     )
     others: list[str] = Field(
@@ -120,19 +120,19 @@ class StreamsBootstrapValues(HelmAppValues):
         description=describe_attr("image", __doc__),
     )
 
-    image_tag: str = Field(
-        default="latest",
+    image_tag: str | None = Field(
+        default=None,
         pattern=IMAGE_TAG_PATTERN,
         description=describe_attr("image_tag", __doc__),
     )
 
-    image_pull_policy: ImagePullPolicy = Field(
-        default=ImagePullPolicy.ALWAYS,
+    image_pull_policy: ImagePullPolicy | None = Field(
+        default=None,
         description=describe_attr("image_pull_policy", __doc__),
     )
 
-    image_pull_secrets: list[dict[str, str]] = Field(
-        default_factory=list,
+    image_pull_secrets: list[dict[str, str]] | None = Field(
+        default=None,
         description=describe_attr("image_pull_secret", __doc__),
     )
 
@@ -140,91 +140,88 @@ class StreamsBootstrapValues(HelmAppValues):
         description=describe_attr("kafka", __doc__),
     )
 
-    resources: Resources = Field(
-        default=Resources(
-            requests=ResourceDefinition(cpu="100m", memory="500Mi"),
-            limits=ResourceDefinition(cpu="300m", memory="2G"),
-        ),
+    resources: Resources | None = Field(
+        default=None,
         description=describe_attr("resources", __doc__),
     )
 
-    ports: list[PortConfig] = Field(
-        default_factory=list,
+    ports: list[PortConfig] | None = Field(
+        default=None,
         description=describe_attr("ports", __doc__),
     )
 
-    service: ServiceConfig = Field(
-        default_factory=ServiceConfig,
+    service: ServiceConfig | None = Field(
+        default=None,
         description=describe_attr("service", __doc__),
     )
 
-    configuration_env_prefix: str = Field(
-        default="APP",
+    configuration_env_prefix: str | None = Field(
+        default=None,
         description=describe_attr("configuration_env_prefix", __doc__),
     )
 
-    command_line: dict[str, str | bool | int] = Field(
-        default_factory=dict,
+    command_line: dict[str, str | bool | int] | None = Field(
+        default=None,
         description=describe_attr("command_line", __doc__),
     )
 
-    env: dict[str, str] = Field(
-        default_factory=dict,
+    env: dict[str, str] | None = Field(
+        default=None,
         description=describe_attr("env", __doc__),
     )
 
-    secrets: dict[str, str] = Field(
-        default_factory=dict,
+    secrets: dict[str, str] | None = Field(
+        default=None,
         description=describe_attr("secrets", __doc__),
     )
 
-    secret_refs: dict[str, Any] = Field(
-        default_factory=dict,
+    secret_refs: dict[str, Any] | None = Field(
+        default=None,
         description=describe_attr("secret_refs", __doc__),
     )
 
-    secret_files_refs: list[str] = Field(
-        default_factory=list,
+    secret_files_refs: list[str] | None = Field(
+        default=None,
         description=describe_attr("secret_files_refs", __doc__),
     )
 
-    files: dict[str, Any] = Field(
-        default_factory=dict,
+    files: dict[str, Any] | None = Field(
+        default=None,
         description=describe_attr("files", __doc__),
     )
 
-    java_options: JavaOptions = Field(
-        default_factory=JavaOptions,
+    java_options: JavaOptions | None = Field(
+        default=None,
         description=describe_attr("java_options", __doc__),
     )
 
-    pod_annotations: dict[str, str] = Field(
-        default_factory=dict,
+    pod_annotations: dict[str, str] | None = Field(
+        default=None,
         description=describe_attr("pod_annotations", __doc__),
     )
 
-    pod_labels: dict[str, str] = Field(
-        default_factory=dict,
+    pod_labels: dict[str, str] | None = Field(
+        default=None,
         description=describe_attr("pod_labels", __doc__),
     )
 
-    liveness_probe: dict[str, Any] = Field(
-        default_factory=dict,
+    liveness_probe: dict[str, Any] | None = Field(
+        default=None,
         description=describe_attr("liveness_probe", __doc__),
     )
 
-    readiness_probe: dict[str, Any] = Field(
-        default_factory=dict,
+    readiness_probe: dict[str, Any] | None = Field(
+        default=None,
         description=describe_attr("readiness_probe", __doc__),
     )
 
-    affinity: dict[str, Any] = Field(
-        default_factory=dict,
+    affinity: dict[str, Any] | None = Field(
+        default=None,
         description=describe_attr("affinity", __doc__),
     )
 
-    tolerations: list[Toleration] = Field(
-        default_factory=list,
+    tolerations: list[Toleration] | None = Field(
+        default=None,
         description=describe_attr("tolerations", __doc__),
     )