+++ kpops/const/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "6.0.1"
+__version__ = "8.1.4"
 KPOPS = "KPOps"
 KPOPS_MODULE = "kpops."